# backend/app/api/routes/resumes.py
"""Эндпоинты резюме: загрузка, запуск анализа, получение результата."""

from __future__ import annotations

from typing import Any

from fastapi import APIRouter, Depends, File, UploadFile
from sqlalchemy.orm import Session

from backend.app.api.deps import Principal, require_access, require_resume_owner
from backend.app.api.responses import OWN_ERRORS
from backend.app.core.audit import log_access
from backend.app.core.exceptions import NotFoundError
from backend.app.database.session import get_db
from backend.app.domain import Resume
from backend.app.repositories.resume_repository import ResumeRepository
from backend.app.repositories.user_repository import UserRepository
from backend.app.schemas.resume import ResumeAnalysisOut, ResumeCreate, ResumeOut
from backend.app.services.resume_analysis import ResumeAnalysisService
from backend.app.services.resume_text import extract_resume_text

router = APIRouter(tags=["resumes"])


@router.post(
    "/users/{user_id}/resumes", response_model=ResumeOut, status_code=201, responses=OWN_ERRORS
)
def create_resume(
    user_id: int,
    payload: ResumeCreate,
    db: Session = Depends(get_db),
    principal: Principal = Depends(require_access),
) -> Resume:
    """Загрузка резюме текстом."""
    UserRepository(db).get(user_id)
    log_access(
        action="create",
        resource="resume",
        resource_id=user_id,
        principal_role=principal.role,
        principal_user_id=principal.user_id,
    )
    return ResumeRepository(db).create(user_id, payload.text, payload.filename)


@router.post(
    "/users/{user_id}/resumes/upload",
    response_model=ResumeOut,
    status_code=201,
    responses=OWN_ERRORS,
)
async def upload_resume(
    user_id: int,
    file: UploadFile = File(...),
    db: Session = Depends(get_db),
    principal: Principal = Depends(require_access),
) -> Resume:
    """Загрузка резюме файлом (TXT/MD, PDF или DOCX)."""
    UserRepository(db).get(user_id)
    log_access(
        action="create",
        resource="resume",
        resource_id=user_id,
        principal_role=principal.role,
        principal_user_id=principal.user_id,
        details={"filename": file.filename},
    )
    content = await file.read()
    text = extract_resume_text(file.filename, content)
    return ResumeRepository(db).create(user_id, text, file.filename)


@router.get("/users/{user_id}/resumes", response_model=list[ResumeOut], responses=OWN_ERRORS)
def list_resumes(
    user_id: int,
    db: Session = Depends(get_db),
    principal: Principal = Depends(require_access),
) -> list[Resume]:
    """Резюме пользователя."""
    return ResumeRepository(db).list_by_user(user_id)


@router.get("/resumes/{resume_id}", response_model=ResumeOut, responses=OWN_ERRORS)
def get_resume(
    resume_id: int,
    db: Session = Depends(get_db),
    principal: Principal = Depends(require_resume_owner),
) -> Resume:
    """Резюме по id."""
    return ResumeRepository(db).get(resume_id)


@router.post("/resumes/{resume_id}/analyze", response_model=ResumeAnalysisOut, responses=OWN_ERRORS)
def analyze_resume(
    resume_id: int,
    db: Session = Depends(get_db),
    principal: Principal = Depends(require_resume_owner),
) -> dict[str, Any]:
    """Запуск анализа резюме."""
    log_access(
        action="create",
        resource="resume_analysis",
        resource_id=resume_id,
        principal_role=principal.role,
        principal_user_id=principal.user_id,
    )
    return ResumeAnalysisService(db).analyze(resume_id)


@router.get("/resumes/{resume_id}/analysis", response_model=ResumeAnalysisOut, responses=OWN_ERRORS)
def get_analysis(
    resume_id: int,
    db: Session = Depends(get_db),
    principal: Principal = Depends(require_resume_owner),
) -> dict[str, Any]:
    """Результат анализа резюме."""
    log_access(
        action="read",
        resource="resume_analysis",
        resource_id=resume_id,
        principal_role=principal.role,
        principal_user_id=principal.user_id,
    )
    analysis = ResumeRepository(db).get_analysis(resume_id)
    if analysis is None or not analysis.result:
        raise NotFoundError(f"Анализ для резюме {resume_id} ещё не выполнен")
    return analysis.result