# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillQuest",
    "description": "Игровой бэкенд SkillQuest: диагностика навыков, игровые миссии, рекомендации курсов и стажировок, анализ резюме. Защищённые операции требуют Bearer-токен (см. security и docs/API.md).",
    "contact": {
      "name": "SkillQuest Team",
      "url": "https://github.com/BrainL4g/max-edu-2026"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/skills/{skill_id}": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "Get Skill",
        "description": "Навык по id.",
        "operationId": "get_skill_skills__skill_id__get",
        "parameters": [
          {
            "name": "skill_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Skill Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/users/{user_id}/skills": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "User Skill Map",
        "description": "Skill Map пользователя: уровень и прогресс по каждому навыку.",
        "operationId": "user_skill_map_users__user_id__skills_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillMapItem"
                  },
                  "title": "Response User Skill Map Users  User Id  Skills Get"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{user_id}/resumes": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Create Resume",
        "description": "Загрузка резюме текстом.",
        "operationId": "create_resume_users__user_id__resumes_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "resumes"
        ],
        "summary": "List Resumes",
        "description": "Резюме пользователя.",
        "operationId": "list_resumes_users__user_id__resumes_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumeOut"
                  },
                  "title": "Response List Resumes Users  User Id  Resumes Get"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/resumes/upload": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Upload Resume",
        "description": "Загрузка резюме файлом (TXT/MD, PDF или DOCX).",
        "operationId": "upload_resume_users__user_id__resumes_upload_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume_users__user_id__resumes_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resumes/{resume_id}": {
      "get": {
        "tags": [
          "resumes"
        ],
        "summary": "Get Resume",
        "description": "Резюме по id.",
        "operationId": "get_resume_resumes__resume_id__get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resumes/{resume_id}/analyze": {
      "post": {
        "tags": [
          "resumes"
        ],
        "summary": "Analyze Resume",
        "description": "Запуск анализа резюме.",
        "operationId": "analyze_resume_resumes__resume_id__analyze_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeAnalysisOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/resumes/{resume_id}/analysis": {
      "get": {
        "tags": [
          "resumes"
        ],
        "summary": "Get Analysis",
        "description": "Результат анализа резюме.",
        "operationId": "get_analysis_resumes__resume_id__analysis_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "resume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resume Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeAnalysisOut"
                }
              }
            }
          },
          "401": {
            "description": "Нет или невалидный Bearer-токен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещён: не свой ресурс или не та роль",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Ресурс не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_resume_users__user_id__resumes_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resume_users__user_id__resumes_upload_post"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut",
        "description": "Ошибка API: человекочитаемое сообщение в поле ``detail``."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeAnalysisOut": {
        "properties": {
          "resume_id": {
            "type": "integer",
            "title": "Resume Id"
          },
          "found_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Found Skills"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "direction_match": {
            "type": "number",
            "title": "Direction Match"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "heuristic"
          },
          "ai_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Score"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          }
        },
        "type": "object",
        "required": [
          "resume_id",
          "found_skills",
          "missing_skills",
          "strengths",
          "issues",
          "recommendations",
          "direction_match",
          "summary"
        ],
        "title": "ResumeAnalysisOut",
        "description": "Результат анализа резюме.\n\nПоля ``source``, ``ai_score`` и ``ai_summary`` заполняются, когда включена\nинтеграция с GigaChat; без неё отчёт остаётся эвристическим.",
        "examples": [
          {
            "direction_match": 0.8,
            "found_skills": [
              "Python",
              "SQL"
            ],
            "issues": [
              "Не указан опыт с Docker"
            ],
            "missing_skills": [
              "Docker"
            ],
            "recommendations": [
              "Добавьте раздел про контейнеризацию"
            ],
            "resume_id": 1,
            "strengths": [
              "Опыт разработки на Python"
            ],
            "summary": "Резюме подходит под направление backend на 80%."
          }
        ]
      },
      "ResumeCreate": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ResumeCreate",
        "description": "Создание резюме из текста.",
        "examples": [
          {
            "filename": "cv.txt",
            "text": "Разработчик Python, 2 года опыта..."
          }
        ]
      },
      "ResumeOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "text",
          "uploaded_at"
        ],
        "title": "ResumeOut",
        "description": "Резюме (ответ API).",
        "examples": [
          {
            "filename": "cv.txt",
            "id": 1,
            "text": "Разработчик Python, 2 года опыта...",
            "uploaded_at": "2026-09-29T10:00:00",
            "user_id": 1
          }
        ]
      },
...
      "SkillMapItem": {
        "properties": {
          "skill_id": {
            "type": "integer",
            "title": "Skill Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "experience": {
            "type": "integer",
            "title": "Experience"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "next_level_xp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Level Xp"
          }
        },
        "type": "object",
        "required": [
          "skill_id",
          "name",
          "category",
          "level",
          "experience",
          "progress"
        ],
        "title": "SkillMapItem",
        "description": "Элемент Skill Map пользователя.",
        "examples": [
          {
            "category": "Программирование",
            "experience": 45,
            "level": 2,
            "name": "Python",
            "next_level_xp": 30,
            "progress": 0.75,
            "skill_id": 1
          }
        ]
      },
      "SkillOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category"
        ],
        "title": "SkillOut",
        "description": "Навык (ответ API)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}